refactor: rename index.ts to providerValidator.ts
